add todo for tile art render
