+++ crates/client/src/main.rs
@@ -30,7 +30,9 @@ mod audio;
 mod rendering_helpers;
 mod sprite_constants;
 mod sprite_loader;
-use crate::rendering_helpers::{build_body_ascii, empty_art_12x8_lines_for_position};
+use crate::rendering_helpers::{
+    art_12x8_lines_for_position, build_body_ascii, empty_art_12x8_lines_for_position, block_art_12x8_lines_for_position
+};
 use crate::sprite_constants::{sprite_for_view_reticle, sprite_for_view_reticle_color};
 use crate::sprite_loader::{
     sprite_block_for_fluid, sprite_block_for_spriteref, sprite_block_for_tile, Scale, SpriteLoader,
@@ -1129,88 +1131,6 @@ fn kind_name(kind: ItemKind) -> &'static str {
     }
 }
 
-// Helper: parse #RRGGBB strings into ratatui Color
-fn parse_hex_color(s: &str) -> Option<Color> {
-    let s = s.trim();
-    if let Some(hex) = s.strip_prefix('#') {
-        if hex.len() == 6 {
-            if let (Ok(r), Ok(g), Ok(b)) = (
-                u8::from_str_radix(&hex[0..2], 16),
-                u8::from_str_radix(&hex[2..4], 16),
-                u8::from_str_radix(&hex[4..6], 16),
-            ) {
-                return Some(Color::Rgb(r, g, b));
-            }
-        }
-    }
-    None
-}
-
-// Build 12x8 art lines for entity or dropped item at this position. Returns true if any art was added.
-fn art_12x8_lines_for_position(
-    app: &mut App,
-    pos: lithicrivers_core::components::Position,
-    out: &mut Vec<Line<'static>>,
-) -> bool {
-    // First, try any entity at this position with a SpriteRef
-    for (_e, (e_pos, maybe_sr)) in app
-        .game
-        .world
-        .query::<(
-            &lithicrivers_core::components::Position,
-            Option<&lithicrivers_core::components::SpriteRef>,
-        )>()
-        .iter()
-    {
-        if *e_pos != pos {
-            continue;
-        }
-        if let Some(sr) = maybe_sr {
-            let sd = app.sprite_loader.load_by_spriteref(sr);
-            let color = parse_hex_color(&sd.color);
-            if let Some(block) = sd.art12x8_sprites.first() {
-                for row in block.split('\n') {
-                    let span = match color {
-                        Some(c) => Span::styled(row.to_string(), Style::default().fg(c)),
-                        None => Span::raw(row.to_string()),
-                    };
-                    out.push(Line::from(span));
-                }
-                return true;
-            }
-        }
-    }
-
-    // If there is any entity at this position but without SpriteRef, render a generic entity sprite
-    let mut any_entity = false;
-    for (_e, (e_pos,)) in app
-        .game
-        .world
-        .query::<(&lithicrivers_core::components::Position,)>()
-        .iter()
-    {
-        if *e_pos == pos {
-            any_entity = true;
-            break;
-        }
-    }
-    if any_entity {
-        let sd = app.sprite_loader.load_sprite("entity_generic", "entities");
-        let color = parse_hex_color(&sd.color);
-        if let Some(block) = sd.art12x8_sprites.first() {
-            for row in block.split('\n') {
-                let span = match color {
-                    Some(c) => Span::styled(row.to_string(), Style::default().fg(c)),
-                    None => Span::raw(row.to_string()),
-                };
-                out.push(Line::from(span));
-            }
-            return true;
-        }
-    }
-    false
-}
-
 fn render_look_panel(f: &mut Frame, app: &mut App, area: Rect) {
     let pos = app.look_cursor;
 
@@ -1221,6 +1141,7 @@ fn render_look_panel(f: &mut Frame, app: &mut App, area: Rect) {
         pos.x, pos.y, pos.z
     ))));
 
+    lines.push(Line::from("Entity or item art:"));
     // Gather 12x8 art for entity/dropped item under cursor (if any)
     if art_12x8_lines_for_position(app, pos, &mut lines) {
         lines.push(Line::from(""));
@@ -1232,6 +1153,14 @@ fn render_look_panel(f: &mut Frame, app: &mut App, area: Rect) {
     // Tile info
     let tile_kind = app.game.res.world.get_tile(pos.x, pos.y);
     lines.push(Line::from(Span::raw(format!("Tile: {:?}", tile_kind))));
+
+    lines.push(Line::from("Tile art:"));
+    if block_art_12x8_lines_for_position(app, pos, &mut lines) {
+        lines.push(Line::from(""));
+    } else {
+        lines.extend(empty_art_12x8_lines_for_position());
+        lines.push(Line::from(""));
+    }
 
     // Fluid info
     if let Some(fluid) = app.game.res.fluids.get_fluid(pos) {
